Use assertj; specify yaml language and move openFile to Rule
